browser-use/requirements.txt
langchain-ollama>=0.3.0
langchain-openai>=0.3.11
browser-use>=0.1.48
python-dotenv>=1.1.0
